add autoSave property
The autoSave property will persist every CodeMirror change to the
underlying text area by repeatedly calling codemirror.save().

This fixes an issue where a user will submit the form and expect the
current codemirror text to submit as the textarea value but due to some
quirk or unclear behavior, the textarea value is not set by the time an
onSubmit handler is triggered.